# api/rest/swagger.yaml
openapi: "3.0.3"
info:
  title: Goph Profile API
  description: API для управления аватарами пользователей
  version: 1.0.0
servers:
  - url: /api/v1
    description: API сервер
...
paths:
...
  /avatars/{avatar_id}:
    get:
      tags:
        - avatars
      description: Получение аватарки по ID
      operationId: getAvatar
      parameters:
        - name: avatar_id
          in: path
          required: true
          description: Идентификатор аватара
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: size
          in: query
          required: false
          description: Размер аватарки
          schema:
            type: string
            enum:
              - "100x100"
              - "300x300"
              - "original"
            default: "original"
            example: "300x300"
        - name: format
          in: query
          required: false
          description: Формат изображения
          schema:
            type: string
            enum:
              - jpeg
              - png
              - webp
            example: "webp"
      responses:
        "200":
          description: Аватар успешно получен
          headers:
            Content-Type:
              description: Тип контента изображения
              schema:
                type: string
                enum:
                  - image/jpeg
                  - image/png
                  - image/webp
            Cache-Control:
              description: Заголовок кеширования
              schema:
                type: string
                example: "max-age=86400"
            ETag:
              description: Хеш для кеширования
              schema:
                type: string
                example: "\"abc123\""
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          description: Аватар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Avatar not found"
    delete:
      tags:
        - avatars
      description: Удаление аватара по ID
      operationId: deleteAvatar
      parameters:
        - name: avatar_id
          in: path
          required: true
          description: Идентификатор аватара
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: X-User-ID
          in: header
          required: true
          description: Идентификатор пользователя
          schema:
            type: string
            example: "user-123"
      responses:
        "204":
          description: Аватар успешно удален
        "403":
          description: Доступ запрещен — пользователь может удалять только свои аватары
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Доступ запрещён. Можно удалять только свои аватары"
        "404":
          description: Аватар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Avatar not found"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "Неверный формат файла"
...